fastq_strand.py: relocate from 'QC-pipeline' to bcftbx/cli.
